# [FIX] web, mail: speed up read_progress_bar