--- a/addons/web/models/models.py
+++ b/addons/web/models/models.py
@@ -1,11 +1,19 @@
 # -*- coding: utf-8 -*-
-from datetime import datetime
 import babel.dates
 import pytz
 
-from odoo.tools import pycompat, date_utils
-from odoo.tools.misc import DEFAULT_SERVER_DATETIME_FORMAT, DEFAULT_SERVER_DATE_FORMAT
-from odoo import _, api, fields, models
+from odoo import api, fields, models
+from odoo.exceptions import UserError
+from odoo.tools import date_utils
+
+
+DISPLAY_DATE_FORMATS = {
+    'day': 'dd MMM yyyy',
+    'week': "'W'w YYYY",
+    'month': 'MMMM yyyy',
+    'quarter': 'QQQ yyyy',
+    'year': 'yyyy',
+}
 
 
 class Base(models.AbstractModel):
@@ -25,24 +33,47 @@ class Base(models.AbstractModel):
         :return a dictionnary mapping group_by values to dictionnaries mapping
                 progress bar field values to the related number of records
         """
+        group_by_fname = group_by.partition(':')[0]
+        field_type = self._fields[group_by_fname].type
+        if field_type == 'selection':
+            selection_labels = dict(self.fields_get()[group_by]['selection'])
+
+        def adapt(value):
+            if field_type == 'selection':
+                value = selection_labels.get(value, False)
+            if type(value) == tuple:
+                value = value[1]  # FIXME should use technical value (0)
+            return value
+
+        result = {}
+        for group in self._read_progress_bar(domain, group_by, progress_bar):
+            group_by_value = str(adapt(group[group_by]))
+            field_value = group[progress_bar['field']]
+            if group_by_value not in result:
+                result[group_by_value] = dict.fromkeys(progress_bar['colors'], 0)
+            if field_value in result[group_by_value]:
+                result[group_by_value][field_value] += group['__count']
+        return result
+
+    def _read_progress_bar(self, domain, group_by, progress_bar):
+        """ Implementation of read_progress_bar() that returns results in the
+            format of read_group().
+        """
+        try:
+            fname = progress_bar['field']
+            return self.read_group(domain, [fname], [group_by, fname], lazy=False)
+        except UserError:
+            # possibly failed because of grouping on or aggregating non-stored
+            # field; fallback on alternative implementation
+            pass
 
         # Workaround to match read_group's infrastructure
         # TO DO in master: harmonize this function and readgroup to allow factorization
         group_by_modifier = group_by.partition(':')[2] or 'month'
         group_by = group_by.partition(':')[0]
-        display_date_formats = {
-            'day': 'dd MMM yyyy',
-            'week': "'W'w YYYY",
-            'month': 'MMMM yyyy',
-            'quarter': 'QQQ yyyy',
-            'year': 'yyyy'}
 
         records_values = self.search_read(domain or [], [progress_bar['field'], group_by])
-
-        data = {}
         field_type = self._fields[group_by].type
-        if field_type == 'selection':
-            selection_labels = dict(self.fields_get()[group_by]['selection'])
 
         for record_values in records_values:
             group_by_value = record_values[group_by]
@@ -56,27 +87,14 @@ class Base(models.AbstractModel):
                 if field_type == 'datetime' and self._context.get('tz') in pytz.all_timezones:
                     tz_info = self._context.get('tz')
                     group_by_value = babel.dates.format_datetime(
-                        group_by_value, format=display_date_formats[group_by_modifier],
+                        group_by_value, format=DISPLAY_DATE_FORMATS[group_by_modifier],
                         tzinfo=tz_info, locale=locale)
                 else:
                     group_by_value = babel.dates.format_date(
-                        group_by_value, format=display_date_formats[group_by_modifier],
+                        group_by_value, format=DISPLAY_DATE_FORMATS[group_by_modifier],
                         locale=locale)
+                record_values[group_by] = group_by_value
 
-            if field_type == 'selection':
-                group_by_value = selection_labels[group_by_value] \
-                    if group_by_value in selection_labels else False
+            record_values['__count'] = 1
 
-            if type(group_by_value) == tuple:
-                group_by_value = group_by_value[1] # FIXME should use technical value (0)
-
-            if group_by_value not in data:
-                data[group_by_value] = {}
-                for key in progress_bar['colors']:
-                    data[group_by_value][key] = 0
-
-            field_value = record_values[progress_bar['field']]
-            if field_value in data[group_by_value]:
-                data[group_by_value][field_value] += 1
-
-        return data
+        return records_values