The method is used to get progress per column in kanban
view (green-yellow-red-red lines in Project, CRM etc). There are two main
usage:

1. get statistics for ``kanban_state`` (red/green circles)
2. get statistics for ``activity_state`` (colored clock icon for overdue/today/planned)

Before this commit all cases were handled by calling search_read and then
counting records per group in a python script. This is very inefficient,
especially for ``activity_state``.

With the new implementation, we just call ``read_group`` if both grouping
fields (kanban column and progressbar field) are stored (case n.1), or use
patched read_group to make grouping on a computed field ``activity_state``.

1. Performance test on 60 K project.task records (kanban_state):

With a filter for 6 records:

```
| measurement        | before | after |
|--------------------+--------+-------|
| number of queries  |      6 |     4 |
| query time, ms     |      6 |     5 |
| remaining time, ms |      9 |     6 |
```

All records:
```
| measurement        | before | after |
|--------------------+--------+-------|
| number of queries  |     66 |     4 |
| query time, ms     |    580 |    53 |
| remaining time, ms |   1900 |     8 |
```

2. Performance test on 29 K crm.lead records (activity_state):

With a filter for 10 records:

```
| measurement        | before | after |
|--------------------+--------+-------|
| number of queries  |     22 |     4 |
| query time, ms     |     11 |     4 |
| remaining time, ms |     35 |     8 |
```

All records:

```
| measurement        | before | after |
|--------------------+--------+-------|
| number of queries  |   1301 |     4 |
| query time, ms     |   1790 |   382 |
| remaining time, ms |  40000 |     8 |
```

---

opw-2346901
task-1915411