@yelizariev 2021-03-08
Does all that mean that I should drop/rewrite mail activity tests that checks result depending on user in the activity records?

https://github.com/odoo/odoo/blob/62c849521ef06aeff212cbdd7179a49155276233/addons/test_mail/tests/test_mail_activity.py#L359-L372