@odony 2021-03-08
> Does all that mean that I should drop/rewrite mail activity tests that checks result depending on user in the activity records?

It's a question we need to discuss with @std-odoo and @tde-banana-odoo. It seems they have been trying to implement a "per-assignee" point of view for activity states instead of a natural "per-viewer" one, and that has lead to extra complexity in the code. So we need to see how much of that is an actual specification, and how much is just coincidence that was turned into a spec because of backwards-compatibility.

For me the "per-assignee" approaches only has drawbacks (different from the usual behavior of both dates *and* datetime fields in Odoo, difficult to explain to users, unreliable behavior for multi-tz companies due to working hours, more complex implementation, ...). So I would stay away from it because it doesn't seem to implement anything that users might actually rely on. If this behavior was initially introduced without thinking much about it, it seems it would not be a great loss.

In practice, this change should only impact users who are looking at the activities of TZ-distant users, and who actually care about a few hours of difference between `today` and `planned` for those distant users (despite being a totally arbitrary and unreliable metric, because no notion of working hours whatsoever). I have a hard time finding a credible use case for that.

But I could definitely be wrong, it's just my gut feeling.