@yelizariev 2021-03-17
Can't we leave it as is and make functional update later when it will be decided? The discussed functional change doesn't save much time, while the performance update of this PR does.