@Julien00859 2021-03-23
There are many local variables in that function. I understand the point is to name everything but my little mind cannot keep track of so many references.