@rco-odoo 2021-03-24
@yelizariev @odony following a discussion with Olivier, I refactored `_read_progress_bar()` such that it always returns a valid result: it uses `read_group()` when possible, and the naive algorithm otherwise.  This makes the roles of the methods clearer.
I also added some basic test on `read_progress_bar()`.
Besides, I have reworked the override of `_read_progress_bar()` for `activity_state` following some code written by @nseinlet on our production database.  Unless I made a mistake, it is equivalent to Ivan's proposal, but more compact.