@yelizariev 2021-03-24
FYI, progressbar is also covered by UI tests

https://github.com/odoo/odoo/blob/6b332a8b8b8ab230fb1b49d24dfe4c0ebab7daa9/addons/web/static/tests/fields/basic_fields_tests.js#L5683

https://github.com/odoo/odoo/blob/6b332a8b8b8ab230fb1b49d24dfe4c0ebab7daa9/addons/web/static/tests/views/kanban_tests.js#L4716-L4774