@yelizariev 2021-05-06
I've fixed conflicts with [recent commit](https://github.com/odoo-dev/odoo/commit/a3a52ebba4384a6ff593386d74bdee885251cfdc#) and have added a similar test for call  (see file `addons/test_mail/tests/test_read_progress_bar.py` )

``read_progress_bar(domain, group_by='date:week', progress_bar={'field': 'activity_state', ....})`` 

Do we want to keep current commits set? Shall I squash some or all of them?