@yelizariev 2021-06-10
Can we proceed?