@tde-banana-odoo 2021-06-10
Hello,

Concerning mail: I don't like that kind of "some custom SQL / data marshmalling" blobs, especially that there is not any comment or explanation of what you are doing or on how it is compliant with other ACLs / custom rules applied on activities. We already have a lot of custom code in mail that becomes quickly impossible to understand and that we keep because "who knows what is the purpose of that code"... so I hope this one is really interesting.

Moreover concerning mail, commit message (``The code is more compact and straightforward, and it has been tested on
our production database for several years.``) does not mean anything. Does not explain why, what is gained, how standard code is inefficient, how it is now more efficient, how it returns correct results (due to custom ACLs / check management) ...

For the content itself, I let smart people choose.