@yelizariev 2021-07-05
I've measured performance results again. 
For some reason number of queries before the commit are different from original measures, but it's more or less the same.

RESULT: **new version works worse** for activity_state than original version and than current v12 version when domain finds just few records. In that case the query time is ~100ms, while it should be ~10ms. This change in performance was introduced in https://github.com/odoo-dev/odoo/commit/e5c9c672d49375fd97e68c50933ef04d5aa48049

Details are below. Old measurements are in the tables. New ones are after the tables in BEFORE/AFTER sections.

------------------

Performance test on 60 K project.task records (kanban_state):

With a filter for 6 records:

```
| measurement        | before | after |
|--------------------+--------+-------|
| number of queries  |      6 |     4 |
| query time, ms     |      6 |     5 |
| remaining time, ms |      9 |     6 |
```

BEFORE: 8 0.011 0.021
AFTER: 5 0.007 0.009



All records:
```
| measurement        | before | after |
|--------------------+--------+-------|
| number of queries  |     66 |     4 |
| query time, ms     |    580 |    53 |
| remaining time, ms |   1900 |     8 |
```
BEFORE: 67 0.300 1.780
AFTER: 5 0.055 0.012




Performance test on 29 K crm.lead records (activity_state):

With a filter for 10 records:

```
| measurement        | before | after |
|--------------------+--------+-------|
| number of queries  |     22 |     4 |
| query time, ms     |     11 |     4 |
| remaining time, ms |     35 |     8 |
```
BEFORE: 25 0.012 0.032
AFTER: 5 0.095 0.007 :-1: 



All records:

```
| measurement        | before | after |
|--------------------+--------+-------|
| number of queries  |   1301 |     4 |
| query time, ms     |   1790 |   382 |
| remaining time, ms |  40000 |     8 |
```

BEFORE: 1326 1.739 47.934
AFTER:  5 0.129 0.017